isolate sixel support in a sixel.go and common.go
